requirements.txt
asgiref==3.4.1
boto3==1.24.66
botocore==1.27.66
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
cryptography==36.0.1
dj-database-url==0.5.0
Django==4.0.1
django-cleanup==6.0.0
django-cors-headers==3.13.0
django-crispy-forms==1.13.0
django-environ==0.8.1
django-factory-boy==1.0.0
django-heroku==0.3.1
django-jsonfield-backport==1.0.5
django-rest-framework-multi-slug-field==0.2.0
django-rest-knox==4.2.0
django-rest-passwordreset==1.2.1
django-simple-history==3.0.0
django-storages==1.13.1
djangorestframework==3.13.1
docopt==0.6.2
drf-writable-nested==0.6.3
Faker==13.2.0
gunicorn==20.1.0
idna==3.3
jmespath==1.0.1
Pillow==9.0.0
pipreqs==0.4.11
psycopg2==2.9.3
psycopg2-binary==2.9.3
pycparser==2.21
python-dateutil==2.8.2
python-decouple==3.6
python-dotenv==0.20.0
pytz==2021.3
requests==2.27.1
s3transfer==0.6.0
six==1.16.0
smart-open==6.1.0
sqlparse==0.4.2
tzdata==2021.5
urllib3==1.26.8
whitenoise==6.0.0
yarg==0.1.9
